Add integrate-test
