+++ .travis.yml
@@ -11,6 +11,7 @@ services:
   - docker
 env:
   - GO111MODULE=on
+install: true
 
 # define ci-stage
 script:
@@ -26,7 +27,21 @@ script:
   - go mod vendor && go test ./... -coverprofile=coverage.txt -covermode=atomic
   # integrate-test
   - echo 'start integrate-test'
-  - echo 'hello-world'
+  # start zookeeper registry insecure listen in [:]:2181
+  - docker run -d --network host zookeeper
+  - ROOTDIR=$(pwd)
+  - cd ./test/integrate/dubbo-/go-client && docker build . -t  ci-consumer && cd $ROOTDIR
+  - cd ./test/integrate/dubbo-/go-server && docker build . -t  ci-provider && cd $ROOTDIR
+  - docker run -d --network host ci-provider
+  - docker run -it --network host ci-consumer
+
+  # another registry instance, start it by dep
+  # start etcd registry  insecure listen in [:]:2379
+  #- docker run -d --network host k8s.gcr.io/etcd:3.3.10 etcd
+  # start consul registry insecure listen in [:]:8500
+  #- docker run -d --network host consul
+  # start nacos registry insecure listen in [:]:8848
+  #- docker run -d --network host nacos/nacos-server:latest
 
 after_success:
   - bash <(curl -s https://codecov.io/bash)
