+++ test/integrate/dubbo/go-client/go.mod
@@ -1,0 +1,8 @@
+module github.com/apache/dubbo-go/test/integrate/dubbo/go-client
+
+go 1.14
+
+require (
+	github.com/apache/dubbo-go v1.4.1 // indirect
+	github.com/apache/dubbo-go-hessian2 v1.5.0 // indirect
+)
