+++ test/integrate/dubbo/go-client/log.yml
