+++ test/integrate/dubbo/go-client/version.go
