--- test/integrate/dubbo/go-server/app/go.mod
@@ -1,3 +1,0 @@
-module github.com/apache/dubbo-go/test/integrate/dubbo/go-server/app
-
-go 1.14
