+++ test/integrate/dubbo/go-server/go.mod
@@ -1,0 +1,3 @@
+module github.com/apache/dubbo-go/test/integrate/dubbo/go-server
+
+go 1.14
