+++ test/integrate/dubbo/go-server/log.yml
