+++ test/integrate/dubbo/go-server/server.yml
@@ -8,21 +8,18 @@ application:
   module : "dubbogo user-info server"
   version : "0.0.1"
   owner : "ZX"
-  environment : "release"
+  environment : "dev"
 
 registries :
-  "hangzhouzk":
+  "demoZk":
     protocol: "zookeeper"
     timeout	: "3s"
     address: "127.0.0.1:2181"
-    username: ""
-    password: ""
-
 
 services:
   "UserProvider":
     # 可以指定多个registry，使用逗号隔开;不指定默认向所有注册中心注册
-    registry: "hangzhouzk"
+    registry: "demoZk"
     protocol : "dubbo"
     # 相当于dubbo.xml中的interface
     interface : "com.ikurento.user.UserProvider"
@@ -30,22 +27,20 @@ services:
     warmup: "100"
     cluster: "failover"
     methods:
-      - name: "GetUser"
-        retries: 1
-        loadbalance: "random"
-
+    - name: "GetUser"
+      retries: 1
+      loadbalance: "random"
 
 protocols:
   "dubbo":
-      name: "dubbo"
-  #    ip : "127.0.0.1"
-      port: 20000
+    name: "dubbo"
+    port: 20000
 
 
 protocol_conf:
   dubbo:
     session_number: 700
-    session_timeout: "180s"
+    session_timeout: "20s"
     getty_session_param:
       compress_encoding: false
       tcp_no_delay: true
@@ -58,5 +53,5 @@ protocol_conf:
       tcp_read_timeout: "1s"
       tcp_write_timeout: "5s"
       wait_timeout: "1s"
-      max_msg_len: 1024000
+      max_msg_len: 1024
       session_name: "server"
